# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/ActualWorkingTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/AttendanceTimeOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/TotalWorkingTime.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/daily/ExcessOfStatutoryTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/daily/holidayworktime/HolidayWorkFrameTime.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/daily/overtimework/OverTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/daily/withinworktime/WithinStatutoryTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyperformanceprocessing/repository/ReflectBreakTimeOfDailyDomainServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/CalculateDailyRecordServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/ManageReGetClass.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/daily/actualworktime/KrcdtDayAttendanceTime.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/remainingnumber/specialleave/empinfo/grantremainingdata/SpecialLeaveNumberInfo.java
@@ -19,8 +19,8 @@ public class SpecialLeaveNumberInfo {
 	//残数
 	private SpecialLeaveRemainingNumber remainingNumber;
 	
-	public SpecialLeaveNumberInfo(int dayNumberOfGrant, Integer timeOfGrant,Double dayNumberOfUse, Integer timeOfUse,
-			Double useSavingDays, int numberOverDays, Integer timeOver, Double dayNumberOfRemain, Integer timeOfRemain) {
+	public SpecialLeaveNumberInfo(double dayNumberOfGrant, Integer timeOfGrant,double dayNumberOfUse, Integer timeOfUse,
+			Double useSavingDays, double numberOverDays, Integer timeOver, double dayNumberOfRemain, Integer timeOfRemain) {
 		this.grantNumber = SpecialLeaveGrantNumber.createFromJavaType(dayNumberOfGrant, timeOfGrant);
 		this.usedNumber = SpecialLeaveUsedNumber.createFromJavaType(dayNumberOfUse,timeOfUse,useSavingDays,numberOverDays,timeOver);
 		this.remainingNumber = SpecialLeaveRemainingNumber.createFromJavaType(dayNumberOfRemain, timeOfRemain);
