+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/remainingnumber/specialleave/empinfo/grantremainingdata/grantnumber/SpecialLeaveGrantNumber.java
@@ -18,12 +18,12 @@ public class SpecialLeaveGrantNumber {
 	//時間
 	public Optional<TimeOfGrant> timeOfGrant;
 	
-	private SpecialLeaveGrantNumber(int days, Integer minutes) {
+	private SpecialLeaveGrantNumber(double days, Integer minutes) {
 		this.dayNumberOfGrant = new DayNumberOfGrant(days);
 		this.timeOfGrant = minutes != null ? Optional.of(new TimeOfGrant(minutes)) : Optional.empty();
 	}
 
-	public static SpecialLeaveGrantNumber createFromJavaType(int days, Integer minutes) {
+	public static SpecialLeaveGrantNumber createFromJavaType(double days, Integer minutes) {
 		return new SpecialLeaveGrantNumber(days, minutes);
 	}
